change property type

+++ trunk/MyGUIEngine/src/MyGUI_List.cpp
@@ -943,7 +943,7 @@ namespace MyGUI
 		return ITEM_NONE;
 	}
 
-	size_t List::getOptimalHeight()
+	int List::getOptimalHeight()
 	{
 		return (mCoord.height - _getClientWidget()->getHeight()) + (mItemsInfo.size() * mHeightLine);
 	}
